#11439 Use getLargeUpdateCount if update count exceeds max value

Former-commit-id: d73672885aca3e69d2c46f0b243acd7a712f3b4f

+++ plugins/org.jkiss.dbeaver.ext.wmi/src/org/jkiss/dbeaver/ext/wmi/model/WMIStatement.java
@@ -118,7 +118,7 @@ public class WMIStatement implements DBCStatement {
     }
 
     @Override
-    public int getUpdateRowCount() throws DBCException
+    public long getUpdateRowCount() throws DBCException
     {
         return -1;
     }
